Added : User Profile Photo

+++ JWT-with-Spring-Security/src/main/java/com/spring/jwt/repository/ProfilePhotoRepo.java
@@ -3,5 +3,5 @@ package com.spring.jwt.repository;
 import com.spring.jwt.entity.ProfilePhoto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfilePhotoRepo extends JpaRepository<ProfilePhoto, Integer> {
+public interface ProfilePhotoRepo extends JpaRepository<ProfilePhoto, Long> {
 }
